changed the data type of user input from integers to floats

+++ practice_programs_batch_1/input_two_numbers_print_their_quotient_with_decimal.py
@@ -1,6 +1,6 @@
 #Ask user to input two numbers
-first_number= int(input("Enter first number:"))
-second_number= int(input("Enter second number:"))
+first_number= float(input("Enter first number:"))
+second_number= float(input("Enter second number:"))
 
 #Divide the two numbers
 result= first_number/second_number
